Fix issue with lastManualWeather
lastManualWeather only changes if the current weather condition does
not match up with either sides' lastAutoWeather. As both sides'
lastAutoWeather is an empty string (equivalent to the "None" weather
condition) when two Pokemon without auto-weather abilities are active,
lastManualWeather cannot be set to None once it's manually set to any
other weather condition. In this case the weather will never stay as
"None" upon Pokemon changes until a side's lastAutoWeather changes.

The suggested fix for this issue is to always set lastManualWeather
whenever the current weather is "None."